test_cli: run subprocess with shell==True; use os.linesep
our 'test.pyz' is not a valid win32 application, so we must call it through the COMSPEC
(i.e. cmd.exe) by using shell=True, which will resolve automatically the pyz extension
and use python to run it (hopefully).

also the output of the process contains carriage returns on windows, so let's use os.linesep